+++ Packs/ReversingLabs_A1000/Integrations/ReversingLabsA1000v2/doc_imgs/delete_sample.png
